Update build script

+++ util/build.py
@@ -1,6 +1,9 @@
 PROJECTROOT: str = ""
-LAUNCH: bool = True
-PACKAGE: bool = True
+PROJECTNAME: str = "Mathrunner"
+AUTHOR: str = "Gigabrain Studios"
+LAUNCH: bool = False
+PACKAGE: bool = False
+DEBUG: bool = False # if False, JavaScript will be minified
 
 import os
 import subprocess
@@ -16,7 +19,6 @@ def list_all_files() -> [str]:
 
 
 def generate_js() -> [str]:
-    ## combine all .js files in /src into one file
     content: [str] = []
     scripts: [str] = list_all_files()
     for script in scripts:
@@ -35,6 +37,31 @@ def write_js(content: [str], target: str = "build/bundle.js") -> None:
         file.writelines(content)
 
 
+def generate_html() -> [str]:
+    content: str = '<!DOCTYPE html>\n<html lang="en">\n\t<head>\n\t\t<meta charset="UTF-8">\n'
+    content += '\t\t<meta name="viewport" content="width=device-width, initial-scale=1.0">\n'
+    content += '\t\t<meta name="author" content="' + AUTHOR + '">\n'
+    content += '\t\t<title>' + PROJECTNAME + '</title>\n'
+    content += '\t\t<link rel="stylesheet" href="css/style.css">\n'
+    content += '\t\t<script type="text/javascript" src="lib/phaser.min.js"></script>\n'
+    
+    if DEBUG:
+        scripts: [str] = list_all_files()
+        for script in scripts:
+            if script.endswith(".js"):
+                content += f'\t\t<script type="text/javascript" src="{script}"></script>\n'
+    else:
+        content += '\t\t<script type="text/javascript" src="dist/bundle.js"></script>\n'
+
+    content += '\t</head>\n\t<body>\t\n\t</body>\n</html>'
+    return content
+
+
+def write_html(content: [str], target: str = "index.html"):
+    with open(os.path.join(PROJECTROOT, target), mode="w", encoding="utf-8") as file:
+            file.writelines(content)
+
+
 def zip_files(file_name: str, target_dir: str) -> str:
     # create build folder if not exists
     if not os.path.exists(target_dir):
@@ -44,7 +71,8 @@ def zip_files(file_name: str, target_dir: str) -> str:
         # add index.html to the zip file
         zip_file.write(os.path.join(PROJECTROOT, "index.html"), arcname="index.html")
         # add all files in subfolders to the zip file
-        sub_dirs: [str] = ["assets", "css", "dist", "lib", "loc"]
+        sub_dirs: [str] = ["assets", "css", "lib", "loc"]
+        sub_dirs.append("src" if DEBUG else "dist")
         for sub_dir in sub_dirs:
             for root, dirs, filenames in os.walk(os.path.join(PROJECTROOT, sub_dir)):
                 for filename in filenames:
@@ -53,25 +81,29 @@ def zip_files(file_name: str, target_dir: str) -> str:
 
 
 if __name__ == "__main__":
-    print(f"Building {PROJECTROOT}...")
-    content: [str] = generate_js()
-    write_js(content)
-    print("build/bundle.js generated.")
-    
-    print ("Running webpack...")
-    subprocess.run(["npx", "webpack", "--config", "webpack.config.js"], cwd="./", shell=True)
-    print("Webpack finished.")
+    print(f"Building {PROJECTROOT}...")    
+    write_html(generate_html())
+    print("index.html generated.")
+
+    if not DEBUG:
+        write_js(generate_js())
+        print("build/bundle.js generated.")
+
+        print ("Running webpack...")
+        subprocess.run(["npx", "webpack", "--config", "webpack.config.js"], cwd="./", shell=True)
+        print("Webpack finished.")
 
     if LAUNCH:
         os.system("start http://localhost/" + PROJECTROOT )
         print("Project launched in browser.")
     
     if PACKAGE:
-        output: str = zip_files("Mathrunner.zip", "build");
+        output: str = zip_files(PROJECTNAME + ".zip", "build");
         print(f"Packaging finished: {output}")
     
-    os.remove(os.path.join(PROJECTROOT, "build/bundle.js"))
-    print("build/bundle.js deleted.")
+    if os.path.exists(os.path.join(PROJECTROOT, "build/bundle.js")):
+        os.remove(os.path.join(PROJECTROOT, "build/bundle.js"))
+        print("build/bundle.js deleted.")
 
     if os.path.exists(os.path.join(PROJECTROOT, "build")) and len(os.listdir(os.path.join(PROJECTROOT, "build"))) == 0:
         os.rmdir(os.path.join(PROJECTROOT, "build"))
